backend/requirements.txt
# Dépendances Backend Django

# Django et extensions
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0

# Base de données
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Validation et sérialisation
python-decouple==3.8
pillow==10.1.0
python-magic==0.4.27

# IA et ML
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
nltk==3.8.1
spacy==3.7.2

# Traitement de documents
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.9.0

# Cache et performance
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# Email et notifications
django-anymail==10.2

# Sécurité
django-ratelimit==4.1.0
django-security==0.20.0

# Tests
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Développement
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.17.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0

# Monitoring
django-prometheus==2.3.1

# Documentation API
drf-spectacular==0.26.5
